% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.WatZ.2014.R
\name{calc.WatZ.2014}
\alias{calc.WatZ.2014}
\title{Convert Tanner crab size to weight based on the 2014 assessment (Rugolo and Turnock).}
\usage{
calc.WatZ.2014(z, sex, maturity, male = list(immature = list(a = 0.00016,
  b = 3.136), mature = list(a = 0.00016, b = 3.136)),
  female = list(immature = list(a = 0.00064, b = 2.794), mature = list(a
  = 0.00034, b = 2.956)))
}
\arguments{
\item{z}{: vector of sizes (mm CW)}

\item{sex}{: 'MALE' or 'FEMALE'}

\item{maturity}{: 'IMMATURE' or 'MATURE'}

\item{male}{: list of regression coefficients for immature, mature crab (see details)}

\item{female}{: list of regression coefficients for immature, mature crab (see details)}
}
\value{
a vector of weights corresponding to the input sizes
}
\description{
Power-law functions for Tanner crab weight-at-size (g) by sex and maturity
  based on the 2014 assessment (Rugolo and Turnock's parameters).
}
\details{
'male' and 'female' are lists with named elements 'a' and 'b', whose values
reflect the parameters in the eq: w = a*z^b
}
