% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanGrowth.R
\name{calcMeanGrowth.DonaldsonEtAl1981}
\alias{calcMeanGrowth.DonaldsonEtAl1981}
\title{Calculate mean growth as in Donaldson et al., 1981}
\usage{
calcMeanGrowth.DonaldsonEtAl1981(zBs)
}
\arguments{
\item{zBs}{- size bins}
}
\value{
vector
}
\description{
Function to calculate mean growth as in Donaldson et al., 1981.
}
\details{
See Donaldson et al., 1981.
}
