% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanGrowth.R
\name{calcMeanGrowth.tcsam02}
\alias{calcMeanGrowth.tcsam02}
\title{Calculate mean growth as in TCSAM02}
\usage{
calcMeanGrowth.tcsam02(zBs, params)
}
\arguments{
\item{zBs}{\itemize{
\item size bins
}}

\item{params}{\itemize{
\item list of growth parameters
}}
}
\value{
vector
}
\description{
Function to calculate mean growth as in TCSAM02.
}
\details{
Port of TCSAM02 size transition calculations from ADMB to R
}
