% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calcSizeTransitionMatrix.LinearGrowthIncrement.R
\name{calcSizeTransitionMatrix.LinearGrowthIncrement}
\alias{calcSizeTransitionMatrix.LinearGrowthIncrement}
\title{Calculate a size transition matrix.}
\usage{
calcSizeTransitionMatrix.LinearGrowthIncrement(a = 0.7, b = 0.883118,
  beta = 0.75, maxDZ = 50, coeffs = NULL, sizes = seq(from = 27.5,
  to = 182.5, by = 5), showPlot = TRUE, log = FALSE,
  colors = rainbow(1000))
}
\arguments{
\item{a}{-}

\item{b}{-}

\item{beta}{- gamma distribution scale parameter}

\item{maxDZ}{- max growth increment allowed (truncates potential growth)}

\item{coeffs}{- list with coefficients a, b, beta, maxDZ}

\item{showPlot}{- flag (T/F) to plot matrix}

\item{log}{- flag to plot ln-scale matrix (in conjunction w/ showPlot)}

\item{colors}{- palette to plot matrix with}

\item{bins}{- size bins to include in matrix}
}
\value{
matrix
}
\description{
Function to calculate a size transition matrix
}
\details{
Approximates the gamma cdf using the pdf, as in old TCSAM.
}
