% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanGrowth.R
\name{calcSizeTransitionMtrix.tcsam02}
\alias{calcSizeTransitionMtrix.tcsam02}
\title{Calculate the size transition matrix as in TCSAM02}
\usage{
calcSizeTransitionMtrix.tcsam02(zBs, delZ, params)
}
\arguments{
\item{zBs}{- size bins}

\item{delZ}{- size bin width (mm CW)}

\item{params}{- list of growth parameters}
}
\value{
matrix
}
\description{
Function to calculate the size transition matrix as in TCSAM02.
}
\details{
Port of TCSAM02 size transition calculations from ADMB to R
}
