% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admbFunctions.R
\name{elem_prod}
\alias{elem_prod}
\title{Implemetation of ADMB function "elem_prod"}
\usage{
elem_prod(x, y)
}
\arguments{
\item{x}{\itemize{
\item vector
}}

\item{y}{\itemize{
\item vector
}}
}
\value{
elem_prod(x,y)
}
\description{
Function that implements ADMB function "elem_prod"
}
\details{
R version of ADMB elem_prod. x and y must be the same length, or
one or both must have only 1 element.
}
