% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanGrowth.R
\name{getGrowthParams}
\alias{getGrowthParams}
\title{Retrieve a list of growth parameters}
\usage{
getGrowthParams()
}
\value{
list with elements 'assessment', 'tmb', and 'DonaldsonEtAl1981', each of which 
is a tibble with sex-specific growth parameters from the assessment model, the TMB
analysis, and the Donaldson et al., 1981 study.
}
\description{
Function to retrieve a list of growth parameters.
}
\details{
Port of TCSAM02 size transition calculations from ADMB to R
}
