% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFunctions.R
\name{getScaleForAbundance}
\alias{getScaleForAbundance}
\title{Get multiplicative scale factor from input abundance units to output units}
\usage{
getScaleForAbundance(unitsIn, unitsOut)
}
\arguments{
\item{unitsIn}{\itemize{
\item units for input data
}}

\item{unitsOut}{\itemize{
\item units for output data
}}
}
\value{
multiplicative scale: output = scale * input.
}
\description{
Function to get multiplicative scale factor from input abundance units to output units.
}
\details{
None.
}
