% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFunctions.R
\name{getScaleForBiomass}
\alias{getScaleForBiomass}
\title{Get multiplicative scale factor from input biomass units to output units}
\usage{
getScaleForBiomass(unitsIn, unitsOut)
}
\arguments{
\item{unitsIn}{: units for input data}

\item{unitsOut}{: units for output data}
}
\value{
multiplicative scale: output = scale * input.
}
\description{
Function to get multiplicative scale factor from input biomass units to output units.
}
\details{
None.
}
