% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFunctions.R
\name{inputList_AggregateCatchData}
\alias{inputList_AggregateCatchData}
\title{Create an input list for aggregate catch data}
\usage{
inputList_AggregateCatchData(
  type = c("ABUNDANCE", "BIOMASS"),
  dfr = NULL,
  cv = 0.05,
  minErr = 100,
  optFit = c("BY_X", "BY_XM"),
  likeType = c("NORM2", "NORMAL", "LOGNORMAL"),
  likeWgt = 1,
  unitsIn = ifelse(toupper(type[1]) == "ABUNDANCE", c("ONES", "MILLIONS"), c("KG",
    "THOUSANDS_MT", "MILLIONS_LBS")),
  unitsOut = ifelse(toupper(type[1]) == "ABUNDANCE", c("MILLIONS", "ONES"),
    c("THOUSANDS_MT", "KG", "MILLIONS_LBS"))
)
}
\arguments{
\item{type}{: type of aggregate catch data ("ABUNDANCE" or "BIOMASS")}

\item{dfr}{: the dataframe}

\item{cv}{: the default cv (NULL to use cv's from \code{dfr})}

\item{minErr}{: minimum assumed error in catch data (in 1's or kg, depending on \code{type})}

\item{optFit}{: objective function fitting option (e.g., "BY_X","BY_XM")}

\item{likeType}{: likelihood type ("NORM2", "NORMAL" or "LOGNORMAL")}

\item{likeWgt}{: likelihood multiplier}

\item{unitsIn}{: units for input data (possibilities depend on \code{type})}

\item{unitsOut}{: units for output data (possibilities depend on \code{type})}
}
\value{
a list (see Details).
}
\description{
Function to create an input list for aggregate catch data.
}
\details{
See below:

Output list has elements:
\itemize{
\item{dfr - dataframe}
\item{cv - default cv or dataframe with 'year', 'cv', and 'minErr' as columns (e.g., for fishery data)}
\item{minErr - minimum assumed error in catch data (in 1's)}
\item{optFit - objective function fitting option}
\item{likeType - likelihood type}
\item{likeWgt - likelihood multiplier}
\item{unitsIn - input units}
\item{unitsOut - output units}
}

Possible units are:
\itemize{
\item{type = "ABUNDANCE": "ONES" or "MILLIONS"}
\item{type = "BIOMASS": 'THOUSANDS_MT', 'MILLIONS_LBS', or 'KG'}
}
}
