% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFunctions.R
\name{inputList_EffortData}
\alias{inputList_EffortData}
\title{Create an input list for effort data}
\usage{
inputList_EffortData(
  dfr = NULL,
  avgInterval = "[1992:-1]",
  likeType = c("NORM2", "NORMAL", "LOGNORMAL"),
  likeWgt = 1,
  unitsIn = c("ONES", "MILLIONS"),
  unitsOut = c("MILLIONS", "ONES")
)
}
\arguments{
\item{dfr}{\itemize{
\item effort data dataframe
}}

\item{avgInterval}{\itemize{
\item averaging interval (e.g., "\link{1992:-1}")
}}

\item{likeType}{\itemize{
\item likelihood type ("NORM2","NORMAL",or "LOGNORMAL")
}}

\item{likeWgt}{\itemize{
\item likelihood weight (multiplier)
}}

\item{unitsIn}{\itemize{
\item input effort units ("ONES","MILLIONS")
}}

\item{unitsOut}{\itemize{
\item output effort units ("ONES","MILLIONS")
}}
}
\value{
a list (see Details).
}
\description{
Function to create an input list for effort data.
}
\details{
See below:

The input dataframe \code{dfr} should have the following column names
\itemize{
\item{year}
\item{effort}
}

The output list has elements:
\itemize{
\item{dfr - dataframe with columns 'year' and 'effort'}
\item{avgInterval - averaging interval given in TCSAM02 time block format}
\item{likeType - likelihood type}
\item{likeWgt - likelihood weight}
\item{unitsIn - units for input effort}
\item{unitsOut - units for output effort}
}
}
