% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFunctions.R
\name{inputList_SizeCompsData}
\alias{inputList_SizeCompsData}
\title{Create an input list for size comps data}
\usage{
inputList_SizeCompsData(
  dfrZCs = NULL,
  dfrSSs = NULL,
  cutpts = NULL,
  tail_compression = c(0, 0),
  optFit = c("BY_X", "BY_XM"),
  likeType = c("MULTINOMIAL", "DIRICHLET-MULTINOMIAL"),
  likeWgt = 1,
  unitsIn = c("ONES", "MILLIONS"),
  unitsOut = c("MILLIONS", "ONES")
)
}
\arguments{
\item{dfrZCs}{- size comps dataframe}

\item{dfrSSs}{- sample sizes dataframe}

\item{cutpts}{- vector of cutpoints for size comps}

\item{tail_compression}{- 2-element vector giving tail compression factors}

\item{optFit}{- objective function fitting option (e.g., "BY_X","BY_XM")}

\item{likeType}{- likelihood type ("NORM2", "NORMAL" or "LOGNORMAL")}

\item{likeWgt}{-likelihood multiplier}

\item{unitsIn}{- units for input data ("ONES" or "MILLIONS")}

\item{unitsOut}{- units for output data ("ONES" or "MILLIONS")}
}
\value{
a list (see Details).
}
\description{
Function to create an input list for size comps data.
}
\details{
See below:

Output list has elements:
\itemize{
\item{dfrZCs - size comps dataframe}
\item{dfrSSs - sample sizes dataframe}
\item{cutpts - cutpoints for size comps}
\item{tail_compression - 2-element vector giving tail compression factors}
\item{optFit - objective function fitting option}
\item{likeType - likelihood type}
\item{likeWgt - likelihood multiplier}
\item{unitsIn - input units}
\item{unitsOut - output units}
}
}
