% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBubblePlots.R
\name{makeBubblePlots}
\alias{makeBubblePlots}
\title{Make bubble plots of a matrix or list of matrices}
\usage{
makeBubblePlots(
  lst,
  x = "year",
  x.lbl = x,
  y = "size",
  y.lbl = y,
  z = "value",
  z.lbl = z,
  fill = "sex",
  fill.lbl = fill,
  scaleBy = 1,
  normalize = FALSE,
  xlims = NULL,
  ylims = NULL,
  colors = RColorBrewer::brewer.pal(11, "Spectral"),
  alpha = 0.5,
  showPlot = TRUE
)
}
\arguments{
\item{lst}{\itemize{
\item list or matrix
}}

\item{x}{\itemize{
\item column name for x axis
}}

\item{x.lbl}{\itemize{
\item label for x axis
}}

\item{y}{\itemize{
\item column name for y axis
}}

\item{y.lbl}{\itemize{
\item label for y axis
}}

\item{z}{\itemize{
\item column name for bubble sizes
}}

\item{z.lbl}{\itemize{
\item label for sizes
}}

\item{fill}{\itemize{
\item column name for bubble fill colours
}}

\item{fill.lbl}{\itemize{
\item label for fills
}}

\item{scaleBy}{\itemize{
\item value to scale bubbles by
}}

\item{normalize}{\itemize{
\item flag (T/F) to normalize sums to 1
}}

\item{xlims}{\itemize{
\item x-axis limits (or NULL)
}}

\item{ylims}{\itemize{
\item y-axis limits (or NULL)
}}

\item{colors}{\itemize{
\item color palette to use (default is \code{RColorBrewer::brewer.pal(11,"Spectral")})
}}

\item{alpha}{\itemize{
\item transparency level
}}

\item{showPlot}{\itemize{
\item flag to print plot
}}
}
\description{
Function to plot a matrix or list of matrices using a bubble plot.
}
