% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admbFunctions.R
\name{mfexp}
\alias{mfexp}
\title{Implemetation of ADMB function "mfexp"}
\usage{
mfexp(x)
}
\arguments{
\item{x}{\itemize{
\item vector
}}
}
\value{
exp(x)
}
\description{
Function that implements ADMB function "mfexp"
}
\details{
R version of ADMB mfexp.
}
