% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGammaPDFs.R
\name{pdf_gamma}
\alias{pdf_gamma}
\title{Evaluate a gamma pdf}
\usage{
pdf_gamma(x, alpha, theta)
}
\arguments{
\item{x}{\itemize{
\item vector of values at which to evaluate the pdf
}}

\item{alpha}{\itemize{
\item the gamma pdf shape parameter
}}

\item{theta}{\itemize{
\item the gamma pdf scale parameter
}}
}
\value{
vector of values of the gamma pdf evaluated at x
}
\description{
Function to evaluate a gamma pdf.
}
\details{
The underlying gamma pdf (\link{pdf_gamma}) is defined as
\deqn{pdf(x) = \frac{1}{\Gamma(\alpha) \theta^\alpha} x^(\alpha-1) e^{-\frac{x}{\theta}}}
where \eqn{\alpha} (the shape parameter) is related to the mean \eqn{\mu} by
\deqn{\alpha = \frac{\mu}{\theta}}
}
