% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGammaPDFs.R
\name{pdf_z}
\alias{pdf_z}
\title{Evaluate a gamma pdf for \emph{post-molt size} given pre-molt size, parameterized by the mean post-molt size and scale}
\usage{
pdf_z(zp, z_, mnZ, beta)
}
\arguments{
\item{zp}{: the post-molt size}

\item{z_}{: the pre-molt size}

\item{mnZ}{: the mean post-molt size}

\item{beta}{: the gamma distribution scale parameter}
}
\value{
value of the corresponding value of the gamma distribution
}
\description{
Function to evaluate the pdf for \emph{post-molt size} given pre-molt size, parameterized by the
mean post-molt size and scale.
}
\details{
Converts the post-molt size to molt increment and mean molt increment
and uses \link{pdf_dz} to calculate the corresponding value of the pdf.
}
