% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotSizeTransitionMatrix}
\alias{plotSizeTransitionMatrix}
\title{Plot a growth transition matrix as an image.}
\usage{
plotSizeTransitionMatrix(prGr, log = FALSE)
}
\description{
Function to plot a growth transition matrix as an image.
}
\details{

}

