% Generated by roxygen2 (4.0.2): do not edit by hand
\name{selFcn.DoubleNormal}
\alias{selFcn.DoubleNormal}
\title{Calculate a double normal selectivity function.}
\usage{
selFcn.DoubleNormal(x, params, showPlot = FALSE)
}
\arguments{
\item{x}{- age/size bins at which to compute values}

\item{params}{- list or vector of parameters}

\item{showPlot}{- flag (T/F) to plot components of function}
}
\description{
Function to calculate (and optionally plot the components of) a double normal selectivity function.
}
\details{


TBD!!
}

