% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selFcn.DoubleNormal.R
\name{selFcn.DoubleNormal4a}
\alias{selFcn.DoubleNormal4a}
\title{Calculate a double normal selectivity function with "4a" parameterization}
\usage{
selFcn.DoubleNormal4a(zBs, params, fsZ, showPlot = TRUE)
}
\arguments{
\item{zBs}{\itemize{
\item vector of sizes at which to compute values
}}

\item{params}{\itemize{
\item vector of parameters
}}

\item{fsZ}{\itemize{
\item max possible size on the "shelf"
}}

\item{showPlot}{\itemize{
\item flag (T/F) to plot components of function
}}
}
\description{
Function to calculate (and optionally plot the components of) a double normal selectivity function
with "4a" parameterization.
}
\details{
Should replicate TCSAM02 dblnormal4a selectivity function.

\itemize{
\item{params\link{1}: size at which ascending limb hits 1}
\item{params\link{2}: width of ascending limb}
\item{params\link{3}: scaled size at which descending limb departs from 1}
\item{params\link{4}: width of descending limb}
}
}
