% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admbFunctions.R
\name{square}
\alias{square}
\title{Implemetation of ADMB function "square"}
\usage{
square(x)
}
\arguments{
\item{x}{- vector}
}
\value{
x*x
}
\description{
Function that implements ADMB function "square"
}
\details{
R version of ADMB square.
}
