% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFunctions.R
\name{subForTCSAM}
\alias{subForTCSAM}
\title{Make substitutions for "undetermined"}
\usage{
subForTCSAM(x, str)
}
\arguments{
\item{x}{- character vector on which to make substitutions}

\item{str}{- string to substitute}
}
\value{
modified character vector
}
\description{
Function to make substitutions for "undetermined".
}
\details{
None
}
