% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTCSAM02_AggregateCatchData.R
\name{writeTCSAM02_AggregateCatchData}
\alias{writeTCSAM02_AggregateCatchData}
\title{Write aggregate catch data to a connection in TCSAM02 input format}
\usage{
writeTCSAM02_AggregateCatchData(
  con = stdout(),
  dfr = NULL,
  type = c("ABUNDANCE", "BIOMASS"),
  optFit = c("BY_X", "BY_XM"),
  likeType = c("NORM2", "NORMAL", "LOGNORMAL"),
  likeWgt = 1,
  unitsIn = ifelse(toupper(type) == "ABUNDANCE", c("ONES", "MILLIONS"), c("KG",
    "THOUSANDS_MT", "MILLIONS_LBS", "MT", "LBS")),
  unitsOut = ifelse(toupper(type) == "ABUNDANCE", c("MILLIONS", "ONES"),
    c("THOUSANDS_MT", "MILLIONS_LBS", "MT", "KG", "LBS"))
)
}
\arguments{
\item{con}{\itemize{
\item connection to use (default=stdout)
}}

\item{dfr}{\itemize{
\item aggregate catch dataframe to write to connection
}}

\item{type}{\itemize{
\item catch type ("ABUNDANCE" or "BIOMASS")
}}

\item{optFit}{\itemize{
\item objective function fitting option (e.g., "BY_X","BY_XM")
}}

\item{likeType}{\itemize{
\item likelihood type
}}

\item{likeWgt}{\itemize{
\item likelihood weight (multiplier)
}}

\item{unitsIn}{\itemize{
\item input catch units
}}

\item{unitsOut}{\itemize{
\item output catch units
}}
}
\value{
Invisibly returns the connection \code{con} to allow piping.
}
\description{
Function to write aggregate catch data to a connection in TCSAM02 input format.
}
\details{
The input dataframe should have the following column names
\itemize{
\item{year}
\item{sex}
\item{maturity}
\item{shell condition}
\item{value - value column}
\item{cv - cv column}
}
}
