% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSizeTransitionMatrices.R
\name{calcSizeTransitionMatrices}
\alias{calcSizeTransitionMatrices}
\title{Calculate size transition matrices for Tanner crab}
\usage{
calcSizeTransitionMatrices(
  coeffs = list(MALE = list(a = 0.42577, b = 0.971389, beta = 0.75, type =
    "LinearGrowthIncrement"), FEMALE = list(a = 0.7, b = 0.883118, beta = 0.75, type =
    "LinearGrowthIncrement")),
  sizes = seq(from = 27.5, to = 182.5, by = 5)
)
}
\arguments{
\item{coeffs}{- list of sex-specific lists of growth parameters (a, b, beta)}

\item{sizes}{- vector of sizes at which to evaluate}
}
\value{
list with sex-specific size transition matrices
}
\description{
Function to calculate sex-specific size transition matrices for Tanner crab.
}
\details{
Uses \code{calcSizeTransitionMatrix} to calculate sex-specific matrices
}
