% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSizeTransitionMatrix.R
\name{calcSizeTransitionMatrix}
\alias{calcSizeTransitionMatrix}
\title{Calculate a size transition matrix}
\usage{
calcSizeTransitionMatrix(
  coeffs = list(type = "LinearGrowthIncrement", a = 0.7, b = 0.883118, beta = 0.75,
    maxDZ = 50),
  sizes = seq(from = 25, to = 185, by = 1),
  showPlot = FALSE,
  log = FALSE,
  colors = grDevices::rainbow(1000)
)
}
\arguments{
\item{coeffs}{\itemize{
\item list including function 'type' and parameter values
}}

\item{sizes}{\itemize{
\item vector of sizes at which to calculate the sie transition probabilities
}}

\item{showPlot}{\itemize{
\item flag (T/F) to plot the matrix
}}

\item{log}{\itemize{
\item flag to use ln-scale on plot
}}

\item{colors}{\itemize{
\item palette to use to color the plot
}}
}
\value{
\itemize{
\item NULL
}
}
\description{
Function to calculate a size transition matrix.
}
\details{
If
\itemize{
\item 'type'='LinearGrowthIncrement', calls \code{calcSizeTransitionMatrix.LinearGrowthIncrement}.
}
}
