% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotSizeTransitionMatrix.R
\name{plotSizeTransitionMatrix}
\alias{plotSizeTransitionMatrix}
\title{Plot a growth transition matrix as an image.}
\usage{
plotSizeTransitionMatrix(prGr, log = FALSE)
}
\description{
Function to plot a growth transition matrix as an image.
}
\details{

}

