% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSizeTransitionMatrix.R
\name{plotSizeTransitionMatrix}
\alias{plotSizeTransitionMatrix}
\title{Plot a growth transition matrix as an image.}
\usage{
plotSizeTransitionMatrix(prGr, log = FALSE)
}
\arguments{
\item{prGr}{- growth transition matrix}

\item{log}{- flag to plot ln-scale values}
}
\value{
a ggplot2 object
}
\description{
Function to plot a growth transition matrix as an image.
}
\details{
None.
}
