% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selFcn.DoubleNormal.R
\name{selFcn.DoubleNormal}
\alias{selFcn.DoubleNormal}
\title{Calculate a double normal selectivity function}
\usage{
selFcn.DoubleNormal(x, params, showPlot = FALSE, test = FALSE)
}
\arguments{
\item{x}{\itemize{
\item age/size bins at which to compute values
}}

\item{params}{\itemize{
\item vector of parameters (see \link{details})
}}

\item{showPlot}{\itemize{
\item flag (T/F) to plot components of function
}}

\item{test}{\itemize{
\item flag (T/F) to output list with detailed info (T) rather than just vector of sel function values (F)
}}
}
\description{
Function to calculate (and optionally plot the components of) a double normal selectivity function.
}
\details{
TBD!!

params are:
\itemize{
\item{1: location of ascending peak}
\item{2: logistic scale width of plateau}
\item{3: ascending slope  (NOT log scale; i.e. = exp(p\link{3}) if p\link{3} from SS)}
\item{3: descending slope (NOT log scale; i.e. = exp(p\link{4}) if p\link{4} from SS)}
\item{5: logistic scale value in initial bin; i.e. initial = 1./(1.+exp(-params\link{5}));}
\item{6: logistic scale value in final bin;   i.e. final   = 1./(1.+exp(-params\link{6}));}
}
}
