% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selFcn.DoubleNormal.R
\name{selFcn.DoubleNormal4a}
\alias{selFcn.DoubleNormal4a}
\title{Calculate a double normal selectivity function with "4a" parameterization}
\usage{
selFcn.DoubleNormal4a(zBs, params, fsZ, showPlot = TRUE)
}
\arguments{
\item{zBs}{- vector of sizes at which to compute values}

\item{params}{- list or vector of parameters}

\item{fsZ}{- max possible size on the "shelf"}

\item{showPlot}{- flag (T/F) to plot components of function}
}
\description{
Function to calculate (and optionally plot the components of) a double normal selectivity function 
with "4a" parameterization.
}
\details{
Should replicate TCSAM02 dblnormal4a selectivity function.
}
