% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simGrowth}
\alias{simGrowth}
\title{Simulate individual growth trajectories based on a size transition matrix.}
\usage{
simGrowth(ny = 21, zAtR = 1, M = 0.23, prGr = NULL, grCoeffs = NULL,
  prM2M = NULL, m2mCoeffs = NULL, sizes = seq(from = 25, to = 185, by =
  1), showPlot = TRUE, normalize = FALSE)
}
\description{
Function to simulate growth trajectories based on a size transition matrix.
}
\details{

}

