% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInputFile_FleetData.R
\name{writeInputFile_FleetData}
\alias{writeInputFile_FleetData}
\title{Write fleet data (fishery or survey) to a TCSAM input file}
\usage{
writeInputFile_FleetData(
  con = stdout(),
  fleet = NULL,
  type = c("FISHERY", "SURVEY"),
  closed = NULL,
  lstRC = NULL,
  lstTC = NULL,
  lstDC = NULL,
  lstIC = NULL,
  lstEff = NULL
)
}
\arguments{
\item{con}{: connection to write to}

\item{fleet}{: TCSAM fleet name}

\item{type}{: fleet type ('FISHERY' or 'SURVEY')}

\item{closed}{: vector of years when fishery was closed, if fleet is a fishery}

\item{lstRC}{: list with retained catch information (see Details)}

\item{lstTC}{: list with total catch information (see Details)}

\item{lstDC}{: list with discard catch information (see Details)}

\item{lstIC}{: list with index catch information (see Details)}

\item{lstEff}{: list with effort data information}
}
\value{
Invisibly returns the connection to facilitate piping.
}
\description{
Function to write fleet data to a TCSAM input file.
}
\details{
Writes fleet-specific data file in TCSAM02 format.

The input lists \code{lstRC}, \code{lstTC}, \code{lstDC}, and \code{lstIC}
must have the following structure, if not NULL:
\itemize{
\item{lstAbd - NULL, or the \code{lstAbd} input to \code{\link[=writeTCSAM02_CatchData]{writeTCSAM02_CatchData()}} (see \code{\link[=inputList_AggregateCatchData]{inputList_AggregateCatchData()}})}
\item{lstBio - NULL, or the \code{lstAbd} input to\link[=writeTCSAM02_CatchData]{ writeTCSAM02_CatchData()} (see \code{\link[=inputList_AggregateCatchData]{inputList_AggregateCatchData()}})}
\item{lstZCs - NULL, or the \code{lstAbd} input to \code{\link[=writeTCSAM02_CatchData]{writeTCSAM02_CatchData()}} (see \code{\link[=inputList_SizeCompsData]{inputList_SizeCompsData()}})}
}

\code{lstEff} contains input values passed to \code{\link[=writeTCSAM02_EffortData]{writeTCSAM02_EffortData()}} and must have
the following structure (\code{\link[=inputList_EffortData]{inputList_EffortData()}}), if not NULL:
\itemize{
\item{dfr - dataframe with columns 'year' and 'effort'}
\item{avgInterval - averaging interval given in TCSAM02 time block format}
\item{likeType - likelihood type}
\item{likeWgt - likelihood weight}
\item{unitsIn - units for input effort}
\item{unitsOut - units for output effort}
}
}
\seealso{
\code{\link[=inputList_AggregateCatchData]{inputList_AggregateCatchData()}}, \code{\link[=inputList_SizeCompsData]{inputList_SizeCompsData()}}, \code{\link[=inputList_EffortData]{inputList_EffortData()}}.
}
