% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTCSAM02_EffortData.R
\name{writeTCSAM02_EffortData}
\alias{writeTCSAM02_EffortData}
\title{Write effort data to a connection in TCSAM02 input format}
\usage{
writeTCSAM02_EffortData(
  con = stdout(),
  dfr = NULL,
  closed = NULL,
  avgInterval = "[1992:-1]",
  likeType = c("NORM2", "NORMAL", "LOGNORMAL"),
  likeWgt = 1,
  unitsIn = c("ONES", "MILLIONS"),
  unitsOut = c("ONES", "MILLIONS")
)
}
\arguments{
\item{con}{- connection to use (default=stdout)}

\item{dfr}{- effort data dataframe to write to connection}

\item{closed}{- vector of closure years (or NULL)}

\item{avgInterval}{- averaging interval (e.g., "[1992:-1]")}

\item{likeType}{- likelihood type ("NORM2","NORMAL",or "LOGNORMAL")}

\item{likeWgt}{- likelihood weight (multiplier)}

\item{unitsIn}{- input effort units ("ONES","MILLIONS")}

\item{unitsOut}{- output effort units ("ONES","MILLIONS")}
}
\value{
Invisibly returns the connection \code{con} to allow piping.
}
\description{
Function to write effort data to a connection in TCSAM02 input format.
}
\details{
The input values should be in ONES.

The input dataframe should have the following column names
\itemize{
  \item{year}
  \item{effort}
}
}
