% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTCSAM02_SizeCompsData.R
\name{writeTCSAM02_SizeCompsData}
\alias{writeTCSAM02_SizeCompsData}
\title{Write size composition data to a connection in TCSAM02 input format}
\usage{
writeTCSAM02_SizeCompsData(
  con = stdout(),
  dfrZCs = NULL,
  dfrSSs = NULL,
  cutpts = NULL,
  tail_compression = c(0, 0),
  optFit = c("BY_X", "BY_XM"),
  likeType = c("MULTINOMIAL", "DIRICHLET-MULTINOMIAL"),
  likeWgt = 1,
  unitsIn = c("ONES", "MILLIONS"),
  unitsOut = c("MILLIONS", "ONES")
)
}
\arguments{
\item{con}{- connection to use (default=stdout)}

\item{dfrZCs}{- size comps dataframe to write to connection}

\item{dfrSSs}{- sample size dataframe corresponding to dfrZCs}

\item{cutpts}{- vector of cutpoints that were used to create the size comps}

\item{tail_compression}{- two-element vector of compresssion factors}

\item{optFit}{- objective function fitting option (e.g., "BY_X","BY_XM")}

\item{likeType}{- likelihood type ("MULTINOMIAL" or "DIRICHLET-MULTINOMIAL")}

\item{likeWgt}{- likelihood weight (multiplier)}

\item{unitsIn}{- input catch units ("ONES" or "MILLIONS")}

\item{unitsOut}{- output catch units ("MILLIONS" or "ONES)}
}
\value{
Invisibly returns the connection \code{con} to allow piping.
}
\description{
Function to write size composition data to a connection in TCSAM02 input format.
}
\details{
The input values should be in ONES for abundance or KG for biomass data.

The input dataframes should have the following column names
\itemize{
  \item{year}
  \item{sex}
  \item{maturity}
  \item{shell condition}
  \item{size (dfrZCs only)}
  \item{value (dfrZCs) or ss (dfrSSs)}
}
}
